# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fitnessstracker server.",
        "title": "Fitnesstracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/user/alldetails": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch all details by date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserDetails"
                ],
                "summary": "fetch all details",
                "operationId": "user-alldetails-fetch",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2024-01-01\"",
                        "description": "date for which want to see details",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AllDetails"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
        "/user/profile": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch user profile with Email, FullName, Age, Gender, Height, Weight, HealthGoal, ProfilePhoto",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserDetails"
                ],
                "summary": "fetch user profile details",
                "operationId": "user-profiledetails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AllDetails"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/user/profile/": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "update user profile details like Email, FullName, Age, Gender, Height, Weight, HealthGoal, ProfilePhoto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserProfile"
                ],
                "summary": "update user profile details",
                "operationId": "user-profile-update",
                "parameters": [
                    {
                        "description": "The input for update the user profile details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
        "/user/yearly-caloriesburned-details": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch monthly average of calories burned by date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserDetails"
                ],
                "summary": "fetch monthly average of calories burned",
                "operationId": "user-yearlycaloriesburned-fetch",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2024\"",
                        "description": "year for which want to see details",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.YearlyCaloriesBurned"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/user/yearly-weight-details": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch monthly average of weight by date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserDetails"
                ],
                "summary": "fetch monthly average of weight",
                "operationId": "user-yearlyweight-fetch",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2024\"",
                        "description": "year for which want to see details",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.YearlyWeight"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AllDetails": {
            "type": "object",
            "properties": {
                "exerciseDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Exercise"
                    }
                },
                "mealDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Meal"
                    }
                },
                "waterDetails": {
                    "$ref": "#/definitions/models.Water"
                },
                "weightDetails": {
                    "$ref": "#/definitions/models.Weight"
                }
            }
        },
        "models.Exercise": {
            "type": "object",
            "required": [
                "caloriesBurned",
                "duration",
                "exerciseType"
            ],
            "properties": {
                "caloriesBurned": {
                    "type": "number",
                    "maximum": 20000
                },
                "date": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer",
                    "maximum": 1440
                },
                "exerciseType": {
                    "type": "string",
                    "enum": [
                        "weight_lifting",
                        "walking",
                        "running",
                        "gym",
                        "yoga"
                    ]
                }
            }
        },
...
        "models.Meal": {
            "type": "object",
            "required": [
                "caloriesConsumed",
                "ingredients",
                "mealType"
            ],
            "properties": {
                "caloriesConsumed": {
                    "type": "number",
                    "maximum": 20000,
                    "minimum": 0
                },
                "date": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "string"
                },
                "mealType": {
                    "type": "string",
                    "enum": [
                        "breakfast",
                        "lunch",
                        "snacks",
                        "dinner"
                    ]
                }
            }
        },
...
        "models.Message": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Users": {
            "type": "object",
            "required": [
                "age",
                "email",
                "fullName",
                "gender",
                "healthGoal",
                "height",
                "weight"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "maximum": 130
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ]
                },
                "healthGoal": {
                    "type": "string",
                    "enum": [
                        "weight_loss",
                        "weight_gain",
                        "muscle_building",
                        "maintain_body"
                    ]
                },
                "height": {
                    "type": "number",
                    "maximum": 300,
                    "minimum": 50
                },
                "profilePhoto": {
                    "type": "string"
                },
                "weight": {
                    "type": "number",
                    "maximum": 700,
                    "minimum": 2
                }
            }
        },
...
        "models.Water": {
            "type": "object",
            "required": [
                "waterIntake"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "waterIntake": {
                    "type": "number",
                    "maximum": 20
                }
            }
        },
        "models.Weight": {
            "type": "object",
            "required": [
                "dailyWeight"
            ],
            "properties": {
                "dailyWeight": {
                    "type": "number",
                    "maximum": 700,
                    "minimum": 2
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "models.YearlyCaloriesBurned": {
            "type": "object",
            "required": [
                "averageMonthlyCaloriesBurned",
                "month"
            ],
            "properties": {
                "averageMonthlyCaloriesBurned": {
                    "type": "number"
                },
                "month": {
                    "type": "integer"
                }
            }
        },
        "models.YearlyWeight": {
            "type": "object",
            "required": [
                "averageMonthlyWeight",
                "month"
            ],
            "properties": {
                "averageMonthlyWeight": {
                    "type": "number"
                },
                "month": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}